# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get the authenticated user's profile
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /deposits/initiate:
    post:
      operationId: initiateDeposit
      tags: [deposits]
      summary: Initiate a deposit
      description: >
        Creates a pending deposit record and returns the parameters the
        frontend must POST to PayU to continue the payment flow. The
        deposit remains in `pending` status until a PayU webhook confirms
        the outcome.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInitiateInput"
      responses:
        "201":
          description: Deposit initiated — PayU form parameters ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositInitiateResult"
        "400":
          description: Validation error (e.g. invalid package amount)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/mobile:
    post:
      operationId: submitMobileNumber
      tags: [users]
      summary: Submit mobile number for verification
      description: >
        Records the user's Indian mobile number in E.164 format and transitions
        mobileVerificationStatus from not_started to pending. Idempotent if the
        same number is resubmitted while status is pending. Returns an error if
        the number is already verified or belongs to another account.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitMobileNumberInput"
      responses:
        "200":
          description: Mobile number recorded, verification pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error or invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Mobile number already taken by another account or already verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/mobile/verify:
    post:
      operationId: verifyMobileNumber
      tags: [users]
      summary: Verify mobile number using a phone authentication token
      description: >
        Verifies the user's mobile number by validating a phone authentication
        token. Currently uses a stub implementation; will use Firebase Admin SDK
        after deployment. On success, sets mobileVerificationStatus to verified
        and records mobileVerifiedAt. The phone number extracted from the token
        must match the number previously submitted via POST /users/me/mobile.
        Idempotent if the number is already verified.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyMobileNumberInput"
      responses:
        "200":
          description: Mobile number verified (or was already verified)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Token invalid, phone number mismatch, or mobile not in pending state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/email:
    patch:
      operationId: updateEmail
      tags: [users]
      summary: Set or update email address
      description: >
        Stores the user's email address, normalized to lowercase. Idempotent if
        the identical address is submitted again. Allows correcting a typo before
        the first deposit. Returns an error if the address is already registered
        to another account.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmailInput"
      responses:
        "200":
          description: Email address saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error or invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email address already registered to another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/terms:
    post:
      operationId: acceptTerms
      tags: [users]
      summary: Accept Terms & Conditions
      description: >
        Records the timestamp at which the user accepted the Terms & Conditions.
        Idempotent — subsequent calls return 200 without overwriting the original
        acceptance timestamp. Required before a deposit or withdrawal can be
        initiated.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Terms accepted (or were already accepted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: Opaque server-side session token issued on signup/login.
  schemas:
...
    User:
      type: object
      description: A user account, excluding all credential/session material.
      required: [id, username, name, age, accountStatus, createdAt, mobileVerificationStatus]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        age:
          type: integer
        email:
          type: string
          format: email
          nullable: true
          description: >
            Collected during the user's first deposit flow, not at signup.
            Null until the user provides it. Required before a deposit can
            be initiated.
        accountStatus:
          type: string
          enum: [active, suspended, deactivated]
        createdAt:
          type: string
          format: date-time
        mobileNumber:
          type: string
          nullable: true
          description: >
            The user's registered mobile number in E.164 format (+91XXXXXXXXXX).
            Null until submitted. Required and verified before a deposit or
            withdrawal can be initiated.
        mobileVerificationStatus:
          $ref: "#/components/schemas/MobileVerificationStatus"
        termsAcceptedAt:
          type: string
          format: date-time
          nullable: true
          description: >
            Timestamp when the user accepted the Terms & Conditions. Null until
            accepted. Required before a deposit or withdrawal can be initiated.
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    DepositStatus:
      type: string
      enum: [pending, success, failed]
      description: >
        Lifecycle state of a deposit attempt.
        `pending` — initiated, awaiting PayU webhook.
        `success` — PayU confirmed payment; Play Coins credited.
        `failed` — PayU reported failure or the attempt was abandoned.
    Deposit:
      type: object
      description: >
        A single deposit attempt. One row = one attempt. The deposit
        remains `pending` until a PayU webhook updates it to `success`
        or `failed`.
      required:
        - id
        - userId
        - amount
        - coinsToCredit
        - status
        - merchantOrderId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: integer
          description: Deposit amount in whole rupees.
        coinsToCredit:
          type: integer
          description: >
            Play Coins to be credited when the deposit succeeds.
            Equal to `amount` under the current 1 coin = ₹1 rate.
        status:
          $ref: "#/components/schemas/DepositStatus"
        merchantOrderId:
          type: string
          description: >
            Our unique key sent to PayU as `txnid`. Used to match
            incoming webhooks back to this deposit row.
        mihpayId:
          type: string
          nullable: true
          description: >
            PayU's unique transaction identifier, populated when the
            webhook is received. Used for duplicate webhook detection.
        payuTxnId:
          type: string
          nullable: true
          description: PayU's `txnid` echo from the webhook, stored for reconciliation.
        failureReason:
          type: string
          nullable: true
          description: Populated when status is `failed`. Null otherwise.
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: >
            Timestamp when the deposit transitioned to `success` or `failed`.
            Null while status is `pending`.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DepositInitiateInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          enum: [50, 100, 200, 500, 1000, 2000, 5000]
          description: >
            Deposit amount in whole rupees. Must be one of the fixed
            packages: ₹50, ₹100, ₹200, ₹500, ₹1000, ₹2000, ₹5000.
    PayUFormParams:
      type: object
      description: >
        Parameters the frontend must POST to PayU's payment page to
        continue the payment flow. All fields are PayU-mandated.
        The `hash` is pre-computed server-side and must not be
        recomputed or modified by the client.
      required:
        - key
        - txnid
        - amount
        - productinfo
        - firstname
        - email
        - surl
        - furl
        - hash
      properties:
        key:
          type: string
          description: PayU merchant key.
        txnid:
          type: string
          description: Our `merchantOrderId`, sent to PayU as the transaction ID.
        amount:
          type: string
          description: Amount as a string (PayU requires string format, e.g. "100.00").
        productinfo:
          type: string
          description: Human-readable product description shown on the PayU page.
        firstname:
          type: string
          description: User's first name, shown on the PayU page.
        email:
          type: string
          format: email
          description: User's email address, required by PayU.
        phone:
          type: string
          description: User's phone number (optional for PayU but recommended).
        surl:
          type: string
          format: uri
          description: Success redirect URL PayU will POST to after payment.
        furl:
          type: string
          format: uri
          description: Failure redirect URL PayU will POST to after payment.
        hash:
          type: string
          description: >
            SHA-512 HMAC computed server-side using the PayU salt.
            The client must include this value verbatim in the form POST.
    DepositInitiateResult:
      type: object
      required: [deposit, payuFormParams, paymentUrl]
      properties:
        deposit:
          $ref: "#/components/schemas/Deposit"
        payuFormParams:
          $ref: "#/components/schemas/PayUFormParams"
        paymentUrl:
          type: string
          format: uri
          description: >
            PayU payment page URL the frontend must POST payuFormParams to.
            Determined by PAYU_ENV: test → https://test.payu.in/_payment,
            production → https://secure.payu.in/_payment.
...
    MobileVerificationStatus:
      type: string
      enum: [not_started, pending, verified]
      description: >
        Lifecycle state of a user's mobile number verification.
        not_started — mobile number has not been submitted.
        pending     — mobile number submitted, awaiting verification.
        verified    — mobile number has been verified.
    SubmitMobileNumberInput:
      type: object
      required: [mobileNumber]
      properties:
        mobileNumber:
          type: string
          pattern: "^\\+91[6-9]\\d{9}$"
          description: >
            Indian mobile number in E.164 format. Must start with +91 followed
            by a 10-digit number beginning with 6, 7, 8, or 9
            (e.g. +919876543210).
    VerifyMobileNumberInput:
      type: object
      required: [firebaseIdToken]
      properties:
        firebaseIdToken:
          type: string
          minLength: 1
          description: >
            Phone authentication token returned by the client-side SDK after the
            user completes phone verification. The server validates this token
            using verifyPhoneToken() (Firebase Admin SDK after deployment; stub
            in pre-Firebase builds) to extract and confirm the phone number.
    UpdateEmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: >
            The user's email address. Stored in lowercase. Must be unique across
            all accounts. Required before a deposit can be initiated.
...